Added More Tests
Tests to ensure that we filter out jokes with "Chuck Norris", regaurdless of case. This is done via a `MockJokeService` that will always return a joke with "ChuCk NORriS", and `Jester` now contains a method that checks for this (and if the joke is null or empty). Also had to make `GetJoke()` virtual, so that it could be overridden by the `MockJokeService`.